+++ config/struct.go
@@ -2,21 +2,21 @@ package config
 
 // Define the program config
 type Config struct {
-	Name    string `json:"name"`
-	Version string `json:"version"`
+	Name    string
+	Version string
 	Tax     Tax
-	TaxList []Tax `json:"tax"`
+	TaxList []Tax
 }
 
 // Define the tranche on a specific year
 type Tax struct {
 	Year     int
-	Tranches []Tranche `json:"tranches"`
+	Tranches []Tranche
 }
 
 // Define one of the tranch of tax
 type Tranche struct {
-	Min        int     `json:"min"`
-	Max        int     `json:"max"`
-	Percentage float64 `json:"percentage"`
+	Min        int
+	Max        int
+	Percentage float64
 }