+++ core/struct.go
@@ -1,8 +1,17 @@
 package core
 
+import "corpos-christie/config"
+
 // Result from processing income
 type Result struct {
-	Income    int     //Input income from the user
-	Tax       float64 // Tax to pay from the user
-	Remainder float64 // Value Remain for the user
+	income      int          //Input income from the user
+	tax         float64      // Tax to pay from the user
+	remainder   float64      // Value Remain for the user
+	taxTranches []TaxTranche // List of tax by tranches
+}
+
+// Struct to catch tax capture for each tranche
+type TaxTranche struct {
+	tax     float64        // Tax in € on a tranche for the user
+	tranche config.Tranche // Param of this tranche (Min, Max, Percentage)
 }